requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
configparser==5.0.2
progressbar2==3.55.0
selenium==4.0.0
xvfbwrapper==0.2.9
black
